Struct spmc::Receiver

pub struct Receiver<T: Send> {
    // some fields omitted
}

The receiving side of a SPMC channel.

There may be many of these, and the Receiver itself is Sync, so it can be placed in an Arc, or cloned itself.

Methods

impl<T: Send> Receiver<T>

fn try_recv(&self) -> Result<T, TryRecvError>

Try to receive a message, without blocking.

fn recv(&self) -> Result<T, RecvError>

Receive a message from the channel.

If no message is available, this will block the current thread until a message is sent.